Do not fail grub config generation if `dsutil` is not found for whatever reason.
We can live without the serial port in this edge case. Plus when grub config is
regenerated chances are that dsutil will be available and this will get fixed at
that second conf generation.

This fix is specifically needed by the post-upgrade script of `base-os` when going
from 9.x --> 10 because the middleware package is not run as yet and we thus we do
not have dsutil at our disposal.

Ticket: #17876